fixed portalling issue
