+++ src/components/address-form/index.js
@@ -1,10 +1,7 @@
 import React, { useEffect, useState } from "react"
-import { useForm } from "react-hook-form"
-import { Flex, Text } from "rebass"
-import Select from "../molecules/select"
-import Input from "../molecules/input"
 import { countries } from "../../utils/countries"
-import { isEmpty } from "lodash"
+import Input from "../molecules/input"
+import Select from "../molecules/old-select"
 
 const AddressForm = ({
   form = {},
