+++ src/components/molecules/select/components/clear-indicator.tsx
@@ -1,3 +1,4 @@
+import React from "react"
 import { ClearIndicatorProps, GroupBase } from "react-select"
 import XCircleIcon from "../../../fundamentals/icons/x-circle-icon"
 
@@ -14,13 +15,9 @@ export const ClearIndicator = <
       className="group text-grey-40 h-full flex flex-col items-center justify-end"
     >
       <button>
-        <ScreenReaderMessage />
+        <span className="sr-only">Clear values</span>
         <XCircleIcon size={20} />
       </button>
     </div>
   )
-}
-
-const ScreenReaderMessage = () => {
-  return <span className="sr-only">Clear values</span>
 }
