+++ src/components/molecules/select/components/multi-value-label.tsx
