+++ src/components/templates/address-form.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from "react"
-import Select from "../molecules/select"
-import Input from "../molecules/input"
 import { countries } from "../../utils/countries"
+import Input from "../molecules/input"
+import Select from "../molecules/old-select"
 
 const AddressForm = ({
   form = {},
