+++ src/components/templates/discount-general/index.tsx
@@ -7,7 +7,7 @@ import { ActionType } from "../../molecules/actionables"
 import IconTooltip from "../../molecules/icon-tooltip"
 import InputField from "../../molecules/input"
 import ConnectForm from "../../molecules/nested-form"
-import Select from "../../molecules/select"
+import Select from "../../molecules/old-select"
 import StatusSelector from "../../molecules/status-selector"
 import Textarea from "../../molecules/textarea"
 import BodyCard from "../../organisms/body-card"
