+++ src/components/templates/discount-settings/index.tsx
@@ -6,8 +6,8 @@ import AvailabilityDuration from "../../molecules/availability-duration"
 import IconTooltip from "../../molecules/icon-tooltip"
 import InputField from "../../molecules/input"
 import ConnectForm from "../../molecules/nested-form"
+import Select from "../../molecules/old-select"
 import Section from "../../molecules/section"
-import Select from "../../molecules/select"
 import BodyCard from "../../organisms/body-card"
 import RadioGroup from "../../organisms/radio-group"
