+++ src/domain/discounts/discount-form/sections/general.tsx
@@ -1,11 +1,11 @@
-import React, { useEffect, useState } from "react"
-import { useAdminRegions } from "medusa-react"
-import { Controller } from "react-hook-form"
 import clsx from "clsx"
+import { useAdminRegions } from "medusa-react"
+import React, { useEffect, useState } from "react"
+import { Controller } from "react-hook-form"
 import Checkbox from "../../../../components/atoms/checkbox"
 import IconTooltip from "../../../../components/molecules/icon-tooltip"
 import InputField from "../../../../components/molecules/input"
-import Select from "../../../../components/molecules/select"
+import Select from "../../../../components/molecules/old-select"
 import Textarea from "../../../../components/molecules/textarea"
 import CurrencyInput from "../../../../components/organisms/currency-input"
 import { useDiscountForm } from "../form/discount-form-context"
