+++ src/domain/gift-cards/custom-giftcard.tsx
@@ -4,7 +4,7 @@ import { useForm } from "react-hook-form"
 import Button from "../../components/fundamentals/button"
 import InputField from "../../components/molecules/input"
 import Modal from "../../components/molecules/modal"
-import Select from "../../components/molecules/select"
+import Select from "../../components/molecules/old-select"
 import Textarea from "../../components/molecules/textarea"
 import CurrencyInput from "../../components/organisms/currency-input"
 import useNotification from "../../hooks/use-notification"
