+++ src/domain/gift-cards/manage/sections/information.tsx
@@ -9,7 +9,7 @@ import { Controller } from "react-hook-form"
 import TrashIcon from "../../../../components/fundamentals/icons/trash-icon"
 import UnpublishIcon from "../../../../components/fundamentals/icons/unpublish-icon"
 import Input from "../../../../components/molecules/input"
-import Select from "../../../../components/molecules/select"
+import Select from "../../../../components/molecules/old-select"
 import StatusSelector from "../../../../components/molecules/status-selector"
 import TagInput from "../../../../components/molecules/tag-input"
 import BodyCard from "../../../../components/organisms/body-card"
