+++ src/domain/orders/details/address-form.js
@@ -1,9 +1,7 @@
-import React, { useEffect, useState } from "react"
 import { isEmpty } from "lodash"
-
-import Select from "../../../components/molecules/select"
+import React, { useEffect, useState } from "react"
 import Input from "../../../components/molecules/input"
-
+import Select from "../../../components/molecules/old-select"
 import { countries } from "../../../utils/countries"
 
 const AddressForm = ({
