+++ src/domain/orders/new/components/select-shipping.tsx
@@ -3,7 +3,7 @@ import Button from "../../../../components/fundamentals/button"
 import AlertIcon from "../../../../components/fundamentals/icons/alert-icon"
 import TrashIcon from "../../../../components/fundamentals/icons/trash-icon"
 import { SteppedContext } from "../../../../components/molecules/modal/stepped-modal"
-import Select from "../../../../components/molecules/select"
+import Select from "../../../../components/molecules/old-select"
 import CurrencyInput from "../../../../components/organisms/currency-input"
 import { extractOptionPrice } from "../../../../utils/prices"
