+++ src/domain/orders/new/new-order.tsx
@@ -1,27 +1,21 @@
-import React, { useState, useEffect, useCallback, useContext } from "react"
-import _ from "lodash"
-import { useForm } from "react-hook-form"
-import { debounce } from "lodash"
 import { navigate } from "gatsby"
-import Medusa from "../../../services/api"
-import Button from "../../../components/fundamentals/button"
-import useMedusa from "../../../hooks/use-medusa"
-import Items from "./components/items"
-import ShippingDetails from "./components/shipping-details"
-import Billing from "./components/billing-details"
-import Summary from "./components/summary"
-import Select from "../../../components/molecules/select"
-import { extractOptionPrice, extractUnitPrice } from "../../../utils/prices"
-import { removeNullish } from "../../../utils/remove-nullish"
+import _, { debounce } from "lodash"
+import React, { useCallback, useEffect, useState } from "react"
+import { useForm } from "react-hook-form"
+import { LayeredModalContext } from "../../../components/molecules/modal/layered-modal"
 import SteppedModal, {
   SteppedContext,
 } from "../../../components/molecules/modal/stepped-modal"
-import { LayeredModalContext } from "../../../components/molecules/modal/layered-modal"
-import CurrencyInput from "../../../components/organisms/currency-input"
-import TrashIcon from "../../../components/fundamentals/icons/trash-icon"
-import AlertIcon from "../../../components/fundamentals/icons/alert-icon"
+import useMedusa from "../../../hooks/use-medusa"
+import Medusa from "../../../services/api"
+import { extractUnitPrice } from "../../../utils/prices"
+import { removeNullish } from "../../../utils/remove-nullish"
+import Billing from "./components/billing-details"
+import Items from "./components/items"
 import SelectRegionScreen from "./components/select-region"
 import SelectShippingMethod from "./components/select-shipping"
+import ShippingDetails from "./components/shipping-details"
+import Summary from "./components/summary"
 
 const defaultFormValues = {
   region: null,
