+++ src/domain/settings/currencies.tsx
@@ -2,7 +2,7 @@ import { navigate } from "gatsby"
 import { useAdminStore, useAdminUpdateStore } from "medusa-react"
 import React, { useEffect, useState } from "react"
 import BreadCrumb from "../../components/molecules/breadcrumb"
-import Select from "../../components/molecules/select"
+import Select from "../../components/molecules/old-select"
 import BodyCard from "../../components/organisms/body-card"
 import TwoSplitPane from "../../components/templates/two-split-pane"
 import useNotification from "../../hooks/use-notification"
