+++ src/domain/settings/regions/details.tsx
@@ -11,7 +11,7 @@ import Spinner from "../../../components/atoms/spinner"
 import DuplicateIcon from "../../../components/fundamentals/icons/duplicate-icon"
 import TrashIcon from "../../../components/fundamentals/icons/trash-icon"
 import Input from "../../../components/molecules/input"
-import Select from "../../../components/molecules/select"
+import Select from "../../../components/molecules/old-select"
 import BodyCard from "../../../components/organisms/body-card"
 import CurrencyInput from "../../../components/organisms/currency-input"
 import DeletePrompt from "../../../components/organisms/delete-prompt"
