+++ src/domain/settings/regions/new-shipping.tsx
@@ -9,7 +9,7 @@ import Spinner from "../../../components/atoms/spinner"
 import Button from "../../../components/fundamentals/button"
 import Input from "../../../components/molecules/input"
 import Modal from "../../../components/molecules/modal"
-import Select from "../../../components/molecules/select"
+import Select from "../../../components/molecules/old-select"
 import CurrencyInput from "../../../components/organisms/currency-input"
 import useNotification from "../../../hooks/use-notification"
 import { Option } from "../../../types/shared"
