+++ src/domain/settings/regions/new.tsx
@@ -4,7 +4,7 @@ import { useForm } from "react-hook-form"
 import Button from "../../../components/fundamentals/button"
 import Input from "../../../components/molecules/input"
 import Modal from "../../../components/molecules/modal"
-import Select from "../../../components/molecules/select"
+import Select from "../../../components/molecules/old-select"
 import CurrencyInput from "../../../components/organisms/currency-input"
 import useNotification from "../../../hooks/use-notification"
 import { countries as countryData } from "../../../utils/countries"
